Added BaseGmailApiTest. Refactored code.| #2692
