+++ FlowCrypt/src/androidTest/java/com/flowcrypt/email/ui/gmailapi/StandardReplyAllComposeGmailApiFlow.kt
@@ -34,6 +34,7 @@ import com.flowcrypt.email.rules.RetryRule
 import com.flowcrypt.email.rules.ScreenshotTestRule
 import com.flowcrypt.email.ui.base.BaseComposeGmailFlow
 import com.flowcrypt.email.ui.base.BaseComposeScreenTest
+import com.flowcrypt.email.ui.base.BaseGmailApiTest
 import jakarta.mail.internet.MimeMultipart
 import okhttp3.mockwebserver.Dispatcher
 import okhttp3.mockwebserver.MockResponse
@@ -55,7 +56,7 @@ import org.junit.runner.RunWith
 @OutgoingMessageConfiguration(
   to = [],
   cc = [],
-  bcc = [BaseComposeGmailFlow.DEFAULT_BCC_RECIPIENT],
+  bcc = [BaseGmailApiTest.DEFAULT_BCC_RECIPIENT],
   message = BaseComposeScreenTest.MESSAGE,
   subject = "",
   isNew = false
